PHPCore

+++ Includes/Object/Process/Admin/Template/Delete.process.php
@@ -9,7 +9,7 @@ class Delete extends \Process\ProcessExtend
     /**
      * @var array $require Required data
      */
-    public $require = [
+    public array $require = [
         'data' => [
             'template_name_folder'
         ]
@@ -18,7 +18,7 @@ class Delete extends \Process\ProcessExtend
     /**
      * @var array $options Process options
      */
-    public $options = [];
+    public array $options = [];
 
     /**
      * Body of process
@@ -27,6 +27,10 @@ class Delete extends \Process\ProcessExtend
      */
     public function process()
     {
+        if ($this->data->get('template_name_folder') === 'Default') {
+            return false;
+        }
+
         $file = new File();
         $file->delete(ROOT . '/Styles/' . $this->data->get('template_name_folder'));
